Fixed bugs with starting the autosave timer
Change the load_post_handler to be persistent across the session, so that it is not removed when a new file is loaded. This was causing an issue when opening an existing file as blender opened the main file before loading the selected file resulting in the timer not ever starting.

Added in additional logging in the console, if you believe the autosave feature is not working as expected check the console for [Timed Autosave] entries for the allotted interval.